feat(#520): remove HexData class at all
